server/src/routers/image_processing.py
from fastapi import APIRouter, UploadFile, File
from PIL import Image
import io

router = APIRouter(
    prefix="/image",
    tags=["image-processing"]
)

@router.post("/super-resolution")
async def super_resolution(image: UploadFile = File(...)):
    """
    图像超分处理接口
    """
    # TODO: 实现超分辨率处理逻辑
    return {"message": "Super resolution endpoint - To be implemented"}

@router.post("/cartoonize")
async def cartoonize(image: UploadFile = File(...)):
    """
    图像卡通化处理接口
    """
    # TODO: 实现卡通化处理逻辑
    return {"message": "Cartoonize endpoint - To be implemented"}

@router.post("/style-transfer")
async def style_transfer(
    content_image: UploadFile = File(...),
    style_image: UploadFile = File(...)
):
    """
    图像风格迁移接口
    """
    # TODO: 实现风格迁移处理逻辑
    return {"message": "Style transfer endpoint - To be implemented"}
